# packages/frontend/static/api-spec/openapi.yaml
openapi: 3.1.1
info:
  title: Stüble REST-API
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Logs in.
      description: Logs in and returns the authentication cookie.
      operationId: login
      requestBody:
        required: true
        description: A JSON object containing the username and password.
        content:
          application/json:
            schema:
              $ref: "#/components/requestBodies/LoginRequest"
      responses:
        "200":
          description: >
            Successfully authenticated.
            The session ID is returned in a cookie named `SID`. You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets a cookie named `SID`.
              example: SID=abcde12345; Path=/; HttpOnly
        "401":
          description: Authentication failure.
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
  /auth/signup:
    post:
      tags:
        - auth
      summary: Creates an account.
      description: Creates an account.
      operationId: createAccount
      requestBody:
        required: true
        description: A JSON object containing the user information and privacy policy agreement.
        content:
          application/json:
            schema:
              $ref: "#/components/requestBodies/SignUpRequest"
      responses:
        "204":
          description: >
            Successfully signed up first part. Now email must be verified.
        "400":
          description: Invalid data supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
  /auth/verify_signup:
    post:
      tags:
        - auth
      summary: Verifies account creation.
      description: Verifies an account and returns the authentication cookie.
      operationId: verifiesAccountCreation
      requestBody:
        required: true
        description: A JSON object containing the token from the email.
        content:
          application/json:
            schema:
              $ref: "#/components/requestBodies/SignupVerificationRequest"
      responses:
        "204":
          description: Successfully verified the signup. The session ID is returned in a cookie named `SID`. You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets a cookie named `SID`.
              example: SID=abcde12345; Path=/; HttpOnly
        "400":
          description: Invalid data supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logs out.
      description: Logs the current user out and clears the authentication cookie.
      operationId: signout
      responses:
        "200":
          description: Successfully logged out.
        "401":
          description: Authentication failure.
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
      security:
        - guest_sid: []
  /auth/delete:
    delete:
      tags:
        - auth
      summary: Deletes the account.
      description: Deletes the user account and clears the authentication cookie. This operation doesn't delete data about previous visits.
      operationId: deleteAccount
      responses:
        "200":
          description: Successfully deleted account.
        "401":
          description: Authentication failure.
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
      security:
        - guest_sid: []
...
  /user:
    put:
      tags:
        - users
      summary: Create an user.
      description: Create a new verified user.
      operationId: createUser
      requestBody:
        required: true
        description: A JSON string specifying the user.
        content:
          application/json:
            schema:
              $ref: "#/components/requestBodies/UserCreationRequest"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/GuestIntern"
        "422":
          description: User doesn't live in this student residence.
        "403":
          description: Authorization failure (not a host).
        "401":
          description: Authentication failure.
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
      security:
        - host_sid: []

    post:
      tags:
        - users
      summary: Modify an user.
      description: Modifies properties of a specified user.
      operationId: modifyUser
      requestBody:
        required: true
        description: A JSON object containing values to be modified.
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "common.yaml#/components/schemas/GuestIntern"
                - $ref: "common.yaml#/components/schemas/GuestExtern"
              discriminator: id
              required:
                - id
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Guest"
        "422":
          description: |-
            User doesn't live in this student residence. \
            Checked when the verified property is set to `true`.
        "403":
          description: Authorization failure (not a host).
        "401":
          description: Authentication failure.
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
      security:
        - host_sid: []

    get:
      tags:
        - users
      summary: Get user properties.
      description: Get own user properties.
      operationId: getUser
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/User"
        "401":
          description: Authentication failure.
        "5xx":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
      security:
        - guest_sid: []
...
components:
...
  requestBodies:
    LoginRequest:
      type: object
      properties:
        user:
          type: string
          example: john.smith@example.com
        password:
          type: string
          example: "12345"
    SignUpRequest:
      allOf:
        - $ref: "common.yaml#/components/schemas/User"
        - type: object
          properties:
            username:
              type: string
              example: JohnSmith
            email:
              type: string
              example: john.smith@example.com
            password:
              type: string
              example: "12345"
            privacyPolicy:
              type: boolean
              const: true
    SignupVerificationRequest:
      type: object
      properties:
        token:
          type: string
          description: Token from the verification email
          example: abcd1234
    UserCreationRequest:
      allOf:
        - $ref: "common.yaml#/components/schemas/NameProperties"
        - $ref: "common.yaml#/components/schemas/LocationProperties"
        - type: object
          properties:
            verified:
              type: boolean
              const: true
              description: Identity verified (via identity card or key)
            privacyPolicy:
              type: boolean
              const: true
...
  responses:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    guest_sid:
      description: Session id of a normal guest user
      type: apiKey
      name: SID
      in: cookie
    host_sid:
      description: Session id of a host user (includes guest rights)
      type: apiKey
      name: SID
      in: cookie
    tutor_sid:
      description: Session id of an tutor user (includes host rights)
      type: apiKey
      name: SID
      in: cookie
    admin_sid:
      description: Session id of an admin user (includes tutor rights)
      type: apiKey
      name: SID
      in: cookie